Add detector-element analyzer point calculation to Arm
As a template for the calculation needed in the data reduction, using
only information available from within the NeXus file(s), it is possible
to calculate the analyzer scattering point for each detector element.
This is needed to work out, e.g., kf and the secondary path length per
detector element.

Notably the calculation requires assuming that all crystal scattering
vectors lie in a plane perpendicular to the crystal blade long axis.
This may not be reasonable in general, but hopefully is in this case.
The justification involves assuming isotropic crystal mosaic in an out
of the scattering plane, such that moving tau slightly in the plane does
not move the mean direction of the out-of-plane distribution away from
0.